Updates JS!
I am complete with the web vision that i have in mind. I have the basic JS that takes in the user-registered account and is held in storage locally, and when logged in they'll have access to see the database of fictional people.   Very easy to use and very easy to understand but this project is meant to help me understand javascript