ERROR FIXED!
I'm not sure if it was a typo or I meant to do something, but it appeared the register and log in form was broken, as no script/messages were appearing if the user inputted text into the box. Turns out a single "w" that was in line 13, I learned that something very simple can break everything! besides that, I also fixed the card class to have the login/register card to be consistently centered,  and now I made the credits (my GitHub on the bottom) to disappear after leaving the index